Move test_host_components into directory
