+++ crates/core/tests/test_host_components/mod.rs
@@ -1,0 +1,2 @@
+pub mod ml;
+pub mod multiplier;
